renamed and refactored login_button
